@@ -1,31 +1,28 @@
 {
-  "name": "my-app",
-  "version": "0.1.0",
+  "name": "dbonfs-landy-react-template",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@types/node": "^15.14.9",
     "@types/react": "^17.0.68",
-    "@types/react-dom": "^17.0.21",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.28",
     "antd": "^4.24.14",
     "i18next": "^19.9.2",
     "i18next-browser-languagedetector": "^6.1.8",
     "i18next-xhr-backend": "^3.2.2",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-awesome-reveal": "^3.8.1",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-i18next": "^11.18.6",
     "react-router-dom": "^5.3.4",
-    "react-scripts": "5.0.1",
     "styled-components": "^5.3.11",
     "typescript": "^4.9.5"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "preview": "vite preview"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -41,5 +38,13 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/node": "^22.7.5",
+    "@types/react-dom": "^18.3.1",
+    "@vitejs/plugin-react": "^4.3.2",
+    "vite": "^5.4.8",
+    "vite-plugin-html": "^3.2.2",
+    "vite-plugin-pwa": "^0.20.5"
   }
 }
